pyproject.toml
[tool.poetry]
name = "mesadelivros"
version = "0.1.0"
description = ""
authors = ["Uadson Feitosa <uadsonpy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.5"
django = "^4.1.4"
pillow = "^9.3.0"
python-decouple = "^3.6"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
